Core/DBLayer: Added missing fixes from 395ac461a86f23bbb66d555f528914b28d63042e

+++ src/server/game/Handlers/PetitionsHandler.cpp
@@ -337,7 +337,7 @@ void WorldSession::SendPetitionQueryOpcode(uint64 petitionguid)
         Field* fields = result->Fetch();
         ownerguid = MAKE_NEW_GUID(fields[0].GetUInt32(), 0, HIGHGUID_PLAYER);
         name      = fields[1].GetString();
-        type      = fields[2].GetUInt32();
+        type      = fields[2].GetUInt8();
     }
     else
     {
